Fix interaction between selection and mouseover
The highlighting would be incorrect after deselecting a horse,
because mouseovers were handled using direct DOM manipulation
that was ignored if a horse is highlighted.

By making the mouseover_horse part of the state and handling
all the display logic in update(), this is fixed.

The moral of the story is not to manipulate the DOM except in
the draw() and update() functions.