Avoid rebuilding DOM on resize
There was no need to rebuild the DOM on resize, because nothing in
the DOM depends on the window size anyway.

Also the rebuild code was incorrect, in that it removed the svg but not
the #replay div before calling createDom() and so a new copy of the
resize div was being added on each resize event.